NP-2547 Move URL to form values to helpers
